pyproject.toml
[tool.poetry]
name = "wkiskas-chat"
version = "0.1.0"
description = ""
authors = ["nnicejji"]
readme = "README.md"
packages = [{include = "wkiskas_chat"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
python-jose = "^3.3.0"
websockets = "^10.4"
jinja2 = "^3.1.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
asyncpg = "^0.27.0"
python-multipart = "^0.0.5"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
